requirements.txt
google-generativeai
playwright
websocket-client
websockets
python-dotenv
fastapi
uvicorn
httpx
pillow
pytesseract
opencv-python
numpy